write 페이지 css ㅅ줭
