--- src/main/java/com/venture/suyaho/repository/UserService.java
@@ -1,11 +1,0 @@
-package com.venture.suyaho.repository;
-
-
-import com.venture.suyaho.domain.User;
-import com.venture.suyaho.dto.UserDTO;
-
-
-public interface UserService {
-    User login(UserDTO userDTO);
-    void register(UserDTO userDTO) throws Exception;
-}
