+++ src/main/java/com/venture/suyaho/service/UserService.java
@@ -1,0 +1,11 @@
+package com.venture.suyaho.service;
+
+
+import com.venture.suyaho.domain.User;
+import com.venture.suyaho.dto.UserDTO;
+
+
+public interface UserService {
+    User login(UserDTO userDTO);
+    void register(UserDTO userDTO) throws Exception;
+}
