+++ src/main/resources/templates/trade/lostwrite.html
@@ -32,6 +32,6 @@
     </form>
 </div>
 
-<script src="../../static/write.js"></script>
+<script src="../../static/js/write.js"></script>
 </body>
 </html>
