+++ src/main/resources/templates/trade/posts.html
@@ -7,65 +7,62 @@
     <link rel="stylesheet" th:href="@{/css/style.css}">
     <script type="text/javascript" th:src="@{/js/posts.js}"></script>
 </head>
-<body>    <div th:replace="fragments/header :: header"></div>
-
-
-<a th:href="@{/}">
-    <img src="/img/campus.png" alt="Homepage Image" height="300px" width="100%">
-</a>
-<div class="container">
-    <div class="back">
-        <button id="backButton">뒤로가기</button>
-    </div>
-    <h1>중고거래 목록</h1>
-
-    <div class="search-container">
-        <div class="search-combined">
-            <select class="search-category">
-                <option value="title">제목</option>
-                <option value="name">이름</option>
-            </select>
-            <input type="text" placeholder="검색어를 입력하세요" class="search-input">
-            <button class="search-button">검색</button>
+<body>
+    <div th:replace="fragments/header :: header"></div>
+    <a th:href="@{/}">
+        <img src="/img/campus.png" alt="Homepage Image" height="300px" width="100%">
+    </a>
+    <div class="container">
+        <div class="back">
+            <button id="backButton">뒤로가기</button>
         </div>
-        <form action="write" method="get">
-            <button id="submit-button">작성</button>
-        </form>
-    </div>
+        <h1>중고거래 목록</h1>
 
-    <div class="post-wrapper">
-        <div class="post-container">
-            <table class="post-table">
-                <thead>
-                <tr>
-                    <th>번호</th>
-                    <th>제목</th>
-                    <th>카테고리</th>
-                    <th>등록일시</th>
-                    <th>아이디</th>
-                    <th>거래여부</th>
-                </tr>
-                </thead>
-                <tbody>
-                <tr th:each="trade : ${trade_boards}">
-                    <td th:text="${trade.tradeNum}"></td>
-                    <td th:text="${trade.tradeTitle}"></td>
-                    <td th:text="${trade.tradeCategory}"></td>
-                    <td th:text="${trade.tradeTime}"></td>
-                    <td th:text="${trade.user.userName}"></td> <!-- userName 대신 user.userName 사용 -->
-                    <td th:text="${trade.tradeComplete == 'Y' ? '거래완료' : (trade.tradeComplete == 'I' ? '거래중' : '거래미완료')}"></td>
+        <div class="search-container">
+            <div class="search-combined">
+                <select class="search-category">
+                    <option value="title">제목</option>
+                    <option value="name">이름</option>
+                </select>
+                <input type="text" placeholder="검색어를 입력하세요" class="search-input">
+                <button class="search-button">검색</button>
+            </div>
+            <form action="write" method="get">
+                <button id="submit-button">작성</button>
+            </form>
+        </div>
 
-
-                </tr>
-                </tbody>
-            </table>
+        <div class="post-wrapper">
+            <div class="post-container">
+                <table class="post-table">
+                    <thead>
+                    <tr>
+                        <th>번호</th>
+                        <th>제목</th>
+                        <th>카테고리</th>
+                        <th>등록일시</th>
+                        <th>아이디</th>
+                        <th>거래여부</th>
+                    </tr>
+                    </thead>
+                    <tbody>
+                    <tr th:each="trade : ${trade_boards}">
+                        <td th:text="${trade.tradeNum}"></td>
+                        <td th:text="${trade.tradeTitle}"></td>
+                        <td th:text="${trade.tradeCategory}"></td>
+                        <td th:text="${trade.tradeTime}"></td>
+                        <td th:text="${trade.user.userName}"></td> <!-- userName 대신 user.userName 사용 -->
+                        <td th:text="${trade.tradeComplete == 'Y' ? '거래완료' : (trade.tradeComplete == 'I' ? '거래중' : '거래미완료')}"></td>
+                    </tr>
+                    </tbody>
+                </table>
+            </div>
         </div>
     </div>
-</div>
-<div class="pagination">
-    <button class="prev" id="prev">&#10094;</button>
-    <button class="next" id="next">&#10095;</button>
-</div>
-<div th:replace="fragments/footer :: footer"></div>
+    <div class="pagination">
+        <button class="prev" id="prev">&#10094;</button>
+        <button class="next" id="next">&#10095;</button>
+    </div>
+    <div th:replace="fragments/footer :: footer"></div>
 </body>
 </html>
